requirements.txt
fastapi
uvicorn
bcrypt
PyJWT
sqlalchemy